docs for initialise, moving change_value into instrument_model.py
