+++ tests/test_qml_model_helpers.py
@@ -1,29 +1,6 @@
-from nexus_constructor.qml_models.helpers import change_value, generate_unique_name
+from nexus_constructor.qml_models.helpers import generate_unique_name
 from nexus_constructor.component_type import ComponentType
 from nexus_constructor.component import Component
-
-
-def test_GIVEN_different_attribute_WHEN_change_value_called_THEN_changes_attribute_to_new_value():
-    item = Component(ComponentType.SAMPLE, name="test")
-    change_value(item, "name", "hello")
-    assert item.name == "hello"
-
-
-def test_GIVEN_same_value_WHEN_change_value_called_THEN_does_not_change_attribute():
-    item = Component(ComponentType.SAMPLE, name="test")
-    change_value(item, "name", "test")
-    assert item.name == "test"
-
-
-def test_GIVEN_nonexistent_attr_WHEN_change_value_called_THEN_does_nothing():
-    item = Component(ComponentType.SAMPLE, name="test")
-    attribute_that_shouldnt_exist = "somethingthatdoesntexist"
-    change_value(item, attribute_that_shouldnt_exist, "test")
-    try:
-        getattr(item, attribute_that_shouldnt_exist)
-        assert False
-    except AttributeError:
-        assert True
 
 
 base = "component"
